# demo/eShopModernizing/eShopLegacyMVCSolution/target/catalog-item-management/openapi.yaml
openapi: 3.0.3
info:
  title: eShopModernized Catalog — Catalog Item Management
  version: "1.0"
  description: >
    Server-rendered ASP.NET Core MVC admin UI for catalog items. Endpoints return
    HTML views (GET) or 302 redirects (POST). Documented for contract completeness;
    this is not a JSON API.
servers:
  - url: https://localhost:8080
paths:
...
  /Catalog/Details/{id}:
    get:
      summary: View item details (FLOW-002)
      operationId: details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200": { description: HTML details page }
        "400": { description: Missing id (BR-006/ERR-001) }
        "404": { description: Item not found (BR-007/ERR-002) }
...